Added partial documentation

+++ src/Elpida.Backend.Services/TaskService.cs
@@ -93,8 +93,8 @@ namespace Elpida.Backend.Services
 
 					ResultName = dto.Result.Name,
 					ResultDescription = dto.Result.Description,
-					ResultType = dto.Result.Type,
-					ResultAggregation = dto.Result.Aggregation,
+					ResultType = (int)dto.Result.Type,
+					ResultAggregation = (int)dto.Result.Aggregation,
 					ResultUnit = dto.Result.Unit,
 				}
 			);
